requirements/common.txt
Django
wagtail
wagtailmenus
docutils
django-environ
markdown

pytz
pip
wheel
setuptools
psycopg2

django-compressor>=1.4
django-modelcluster>=0.6
django-taggit>=0.14.0
django-libsass>=0.2
libsass>=0.8.2
Pillow>=3.2.0
html5lib>=0.999999